// G2B4_ReactJS_Project/movie tracker/client/src/model/IDataList.tsx
export default interface IDataList {
    id: string;
    title: string;
    year: string;
    genres: string[];
    ratings: number[];
    poster: string;
    contentRating: string;
    duration: string;
    releaseDate: string;
    averageRating: number;
    originalTitle: string;
    storyline: string;
    actors: string[];
    imdbRating: string;
    posterurl: string;
}